Added: GHOST command on startup
If FurkieBot detects someone is already using the nickname 'FurkieBot',
it will change its own name to FurkieBot_.

After making a connection with the server, it will run a GHOST command,
killing the other person using 'FurkieBot' as nickname. Doing a GHOST
command cannot be done before joining the server as a different name
because you will never be able to reach NickServ to do this.

NOTE: The GHOST command is always run on startup, regardless of whether
the nickname is FurkieBot or not. If you never needed the GHOST command
to begin with, NickServ will simply notify that you can't ghost
yourself.